fix: change ID types and add spring validation starter

+++ src/main/java/org/example/passengerinformationdisplaysystem/departures/LiveDepartureRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LiveDepartureRepository extends JpaRepository<LiveDepartureEntity, Integer> {
+public interface LiveDepartureRepository extends JpaRepository<LiveDepartureEntity, Long> {
 }
